Move types to pages
