+++ package.json
@@ -12,7 +12,7 @@
     "test": "jest --watch",
     "test:ci": "jest --ci",
     "test:e2e": "npx playwright test",
-    "typecheck": "echo 'Typecheck check!' && tsc",
+    "typecheck": "echo 'Typecheck...' && tsc",
     "release": "standard-version"
   },
   "dependencies": {
